Add TLS and Auth to etcd backends
