+++ Makefile
@@ -1,15 +1,11 @@
-ETCD_NODE1 := http://127.0.0.1:4001
-ETCD_NODE2 := http://127.0.0.1:4002
-ETCD_NODE3 := http://127.0.0.1:4003
-ETCD_NODES := ${ETCD_NODE1},${ETCD_NODE2},${ETCD_NODE3}
+ETCD_NODES := https://127.0.0.1:2379
 API_URL := http://localhost:8182
 SERVICE_URL := http://localhost:8181
 PREFIX := /vulcandtest
 SEAL_KEY := 1b727a055500edd9ab826840ce9428dc8bace1c04addc67bbac6b096e25ede4b
-GO15VENDOREXPERIMENT := 1
 
-ETCD_FLAGS := VULCAND_TEST_ETCD_NODES=${ETCD_NODES}
-VULCAN_FLAGS := VULCAND_TEST_ETCD_NODES=${ETCD_NODES} VULCAND_TEST_ETCD_PREFIX=${PREFIX} VULCAND_TEST_API_URL=${API_URL} VULCAND_TEST_SERVICE_URL=${SERVICE_URL} VULCAND_TEST_SEAL_KEY=${SEAL_KEY}
+ETCD_FLAGS := VULCAND_TEST_ETCD_NODES=${ETCD_NODES} VULCAND_TEST_ETCD_USER=root VULCAND_TEST_ETCD_PASS=rootpw VULCAND_TEST_ETCD_TLS=true
+VULCAN_FLAGS := ${ETCD_FLAGS} VULCAND_TEST_ETCD_PREFIX=${PREFIX} VULCAND_TEST_API_URL=${API_URL} VULCAND_TEST_SERVICE_URL=${SERVICE_URL} VULCAND_TEST_SEAL_KEY=${SEAL_KEY}
 
 test: clean
 	go test -v ./... -cover
